refactor: 게임 결과 연산을 Participants에서 BlackjackGame의 책임으로 위임
