+++ src/test/java/domain/participant/ParticipantsTest.java
@@ -69,55 +69,4 @@ class ParticipantsTest {
                 .map(Participant::getCards)
                 .allMatch(cards -> cards.size() == 2);
     }
-
-    @Test
-    void 딜러보다_점수가_높으면_WIN() {
-        // given
-        Participants participants = Participants.of(List.of("배럴"));
-        participants.getDealer().addCard(new Card(Denomination.NINE, Suit.SPADE));
-        List<Participant> players = participants.getPlayers();
-
-        // when
-        players.get(0).addCard(new Card(Denomination.ACE, Suit.SPADE));
-
-        // then
-        assertThat(participants.getResult()).contains(
-                // 딜러 9 vs 배럴 ACE
-                Map.entry("배럴", PlayerGameResult.WIN)
-        );
-    }
-
-    @Test
-    void 딜러와_점수가_같다면_DRAW() {
-        // given
-        Participants participants = Participants.of(List.of("둘리"));
-        participants.getDealer().addCard(new Card(Denomination.NINE, Suit.SPADE));
-        List<Participant> players = participants.getPlayers();
-
-        // when
-        players.get(0).addCard(new Card(Denomination.NINE, Suit.SPADE));
-
-        // then
-        assertThat(participants.getResult()).contains(
-                // 딜러 9 vs 둘리 9
-                Map.entry("둘리", PlayerGameResult.DRAW)
-        );
-    }
-
-    @Test
-    void 딜러보다_점수가_낮으면_LOSE() {
-        // given
-        Participants participants = Participants.of(List.of("패배자이름"));
-        participants.getDealer().addCard(new Card(Denomination.NINE, Suit.SPADE));
-        List<Participant> players = participants.getPlayers();
-
-        // when
-        players.get(0).addCard(new Card(Denomination.TWO, Suit.SPADE));
-
-        // then
-        assertThat(participants.getResult()).contains(
-                // 딜러 9 vs 패배자 1
-                Map.entry("패배자이름", PlayerGameResult.LOSE)
-        );
-    }
 }
